In Game Class:
Finished logic for update method. Should be complete

Added paint method. This method will contain all the method calls for
the Obj.update();

Added the JFrame into the Game class. Can be changed depending on needs
of the game.

Added a start() method, called when the game begins, and repeatedly
calls the update() method.

In Player Class and MyKeyListener Class:

Removed print statements.