Added Room Class
Removed unnecessary Level1 class.

Levels will be created randomly.
Each Map will have multiple rooms that connect to each other.
Each room will fill the game screen, and moving between rooms will
happen through a door in each wall of the room, assuming a room is on
the other side of the door.

Sprites for the player and doors will be required.

We might also include a mini-map.